Fix compiler warnings on printf calls.

+++ int_array.c
@@ -584,7 +584,7 @@ int_dump(NODE *symbol, NODE *ndump)
 		fprintf(output_fp, "flags: %s\n", flags2str(symbol->flags));
 	}
 	indent(indent_level);
-	fprintf(output_fp, "INT_CHAIN_MAX: %lu\n", INT_CHAIN_MAX);
+	fprintf(output_fp, "INT_CHAIN_MAX: %lu\n", (unsigned long) INT_CHAIN_MAX);
 	indent(indent_level);
 	fprintf(output_fp, "array_size: %lu (int)\n", (unsigned long) symbol->array_size);
 	indent(indent_level);
